Fix quantizer registration conflict, add skip tensor copying to convert_to_hf
Root cause: two competing quantizer implementations (hf_integration.py and
hf_quantizer.py) both registered for "helix"/"cdna_v3". Import order determined
which was active. hf_integration.py creates HelixLinear without sidecar buffers,
so HF loader silently skips sidecar keys → PPL=7909 (Coder-3B) / NaN (Qwen-7B).

Fixes:
- Make hf_quantizer.py the sole canonical quantizer (all registrations)
- Remove all @register decorators and alias registrations from hf_integration.py
- Add backward-compat params (codebook_size, sidecar_enabled) to CDNAv3Config
- Add nn.Embedding compression support (codebook/indices → dense reconstruction)
- Register "helix", "hxq", "cdna_v3" aliases from single location

convert_to_hf.py:
- Copy 1D "skip" tensors (A_log, D, dt_bias, norms) from original model
  (compress.py skips these without storing in cdnav3/)
- Add structural validation gate (readable, indices in range, no NaN, key count)
- Add SHA256 and WO-RECEIPT-COST-01 compliant conversion receipts
- Change quant_method from "helix" to "hxq"
- Strip dead SVD factors (killed 2026-03-27)

New tools: batch_gpu_eval.py, ssm_compress_pipeline.py, eval_zamba2_7b_*.py

+++ tools/compress.py
@@ -465,8 +465,12 @@ def compress_model(model_dir: Path, dry_run: bool = False, force: bool = False,
             safe_name = name.replace("/", "_").replace(".", "_")
             exact_path = cdna_dir / f"{safe_name}.npy"
             if not exact_path.exists() or force:
-                tensor_np = source.get_tensor(name)
-                np.save(exact_path, tensor_np)
+                tensor_np = source.get_tensor(name)  # FP32 from get_tensor()
+                # Store at FP16 instead of FP32 to avoid BF16→FP32 inflation.
+                # FP16 has more mantissa bits (10) than BF16 (7), so no precision
+                # loss relative to the original model. Saves 50% for exact tensors
+                # — embed_tokens alone goes from 438 MB to 219 MB on 7B models.
+                np.save(exact_path, tensor_np.astype(np.float16))
                 del tensor_np
             # Write meta.json companion (needed for standalone bias loading)
             meta_path = cdna_dir / f"{safe_name}.npy.meta.json"
